requirements.txt
requests>=2.32.0
python-dotenv>=1.0.1
tabulate>=0.9.0
colorama>=0.4.6
python-binance>=1.0.19